+++ packages/isaacscript-common/src/functions/debug.ts
@@ -1,4 +1,36 @@
+import { ModUpgraded } from "../classes/ModUpgraded";
+import { enableExtraConsoleCommands } from "../features/extraConsoleCommands/exports";
+import { removeFadeIn } from "../features/fadeInRemover";
+import { enableFastReset } from "../features/fastReset";
+import { saveDataManagerSetGlobal } from "../features/saveDataManager/exports";
+import * as logExports from "./log";
 import { log } from "./log";
+import * as logEntitiesExports from "./logEntities";
+
+/**
+ * Helper function to enable some IsaacScript features that are useful when developing a mod. They
+ * should not be enabled when your mod goes to production (i.e. when it is uploaded to the Steam
+ * Workshop).
+ *
+ * The list of development features that are enabled are as follows:
+ *
+ * - `saveDataManagerSetGlobal` - Sets your local variables registered with the save data manager as
+ *   global variables so you can access them from the in-game console.
+ * - `setLogFunctionsGlobal` - Sets the various log functions global so that you can access them
+ *   from the in-game console.
+ * - `enableExtraConsoleCommands` - Enables many extra in-game console commands that make warping
+ *   around easier (like e.g. `angel` to warp to the Angel Room).
+ * - `enableFastReset` - Makes it so that the r key resets the game instantaneously.
+ * - `removeFadeIn` - Removes the slow fade in that occurs at the beginning of the run, so that you
+ *   can immediately start playing or testing.
+ */
+export function enableDevFeatures(mod: ModUpgraded): void {
+  saveDataManagerSetGlobal();
+  setLogFunctionsGlobal();
+  enableExtraConsoleCommands(mod);
+  enableFastReset();
+  removeFadeIn();
+}
 
 /**
  * Helper function to get a stack trace.
@@ -35,6 +67,23 @@ export function isLuaDebugEnabled(): boolean {
   // "package" is not always defined like the Lua definitions imply.
   // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
   return _G.package !== undefined;
+}
+
+/**
+ * Converts every `isaacscript-common` function that begins with "log" to a global function.
+ *
+ * This is useful when printing out variables from the in-game debug console.
+ */
+export function setLogFunctionsGlobal(): void {
+  const globals = _G as Record<string, unknown>;
+
+  for (const [logFuncName, logFunc] of Object.entries(logExports)) {
+    globals[logFuncName] = logFunc;
+  }
+
+  for (const [logFuncName, logFunc] of Object.entries(logEntitiesExports)) {
+    globals[logFuncName] = logFunc;
+  }
 }
 
 /**